Format used in experimental implementation
Some changes to make implementation easier.

The cluster mapping is now 2-level.  We found initialization may take
too long if we need to write a whole cluster mapping.  The 2-level
format can still be converted to qcow2 easily.

The checksum in end pointers can optionally be CRC32 instead of SHA256.
Because we wanted to test other functionality first and we delayed
figuring out how to do RSA and SHA256 on the Microblaze.